refactor(tests): fix failing tests
